# *: Implement RFC-0041 Object Native API