--- a/benches/ops/fs.rs
+++ b/benches/ops/fs.rs
@@ -14,7 +14,6 @@
 use std::env;
 
 use anyhow::Result;
-
 use opendal::services::fs;
 use opendal::Operator;