--- a/src/accessor.rs
+++ b/src/accessor.rs
@@ -15,34 +15,33 @@
 use std::sync::Arc;
 
 use async_trait::async_trait;
-use bitflags::bitflags;
+use futures::AsyncRead;
 
 use crate::error::Result;
-use crate::io::Reader;
-use crate::io::StatefulReader;
+use crate::io::AsyncReadSeek;
+use crate::io::RandomReader;
+use crate::io::SequentialReader;
 use crate::object::Metadata;
 use crate::ops::OpDelete;
-use crate::ops::OpRead;
+use crate::ops::OpRandomRead;
+use crate::ops::OpSequentialRead;
 use crate::ops::OpStat;
-use crate::ops::OpStatefulRead;
 use crate::ops::OpWrite;
+use crate::{BoxedAsyncRead, BoxedAsyncReadSeek};
 
 #[async_trait]
 pub trait Accessor: Send + Sync {
-    fn features(&self) -> Features {
-        unimplemented!()
-    }
     /// Read data from the underlying storage into input writer.
-    async fn read(&self, args: &OpRead) -> Result<Reader> {
+    async fn sequential_read(&self, args: &OpSequentialRead) -> Result<BoxedAsyncRead> {
         let _ = args;
         unimplemented!()
     }
-    async fn stateful_read(&self, args: &OpStatefulRead) -> Result<StatefulReader> {
+    async fn random_read(&self, args: &OpRandomRead) -> Result<BoxedAsyncReadSeek> {
         let _ = args;
         unimplemented!()
     }
     /// Write data from input reader to the underlying storage.
-    async fn write(&self, r: Reader, args: &OpWrite) -> Result<usize> {
+    async fn write(&self, r: BoxedAsyncRead, args: &OpWrite) -> Result<usize> {
         let (_, _) = (r, args);
         unimplemented!()
     }
@@ -67,16 +66,13 @@ pub trait Accessor: Send + Sync {
 /// `Accessor` for `Arc<dyn Accessor>`.
 #[async_trait]
 impl<T: Accessor> Accessor for Arc<T> {
-    fn features(&self) -> Features {
-        self.as_ref().features()
+    async fn sequential_read(&self, args: &OpSequentialRead) -> Result<BoxedAsyncRead> {
+        self.as_ref().sequential_read(args).await
     }
-    async fn read(&self, args: &OpRead) -> Result<Reader> {
-        self.as_ref().read(args).await
+    async fn random_read(&self, args: &OpRandomRead) -> Result<BoxedAsyncReadSeek> {
+        self.as_ref().random_read(args).await
     }
-    async fn stateful_read(&self, args: &OpStatefulRead) -> Result<StatefulReader> {
-        self.as_ref().stateful_read(args).await
-    }
-    async fn write(&self, r: Reader, args: &OpWrite) -> Result<usize> {
+    async fn write(&self, r: BoxedAsyncRead, args: &OpWrite) -> Result<usize> {
         self.as_ref().write(r, args).await
     }
     async fn stat(&self, args: &OpStat) -> Result<Metadata> {
@@ -84,17 +80,5 @@ impl<T: Accessor> Accessor for Arc<T> {
     }
     async fn delete(&self, args: &OpDelete) -> Result<()> {
         self.as_ref().delete(args).await
-    }
-}
-
-bitflags! {
-    /// Features supported by the accessor.
-    #[derive(Default)]
-    pub struct Features: u64 {
-        const READ = 1 << 0;
-        const WRITE = 1 << 1;
-        const STAT = 1 << 2;
-        const DELETE = 1 << 3;
-        const STATEFUL_READ = 1 << 4;
     }
 }