--- a/src/ops.rs
+++ b/src/ops.rs
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-
-
-
 #[derive(Debug, Clone, Default)]
 pub struct OpRead {
     pub path: String,