… Cargo.toml
… examples/basic.rs
… rustfmt.toml
… src/operator.rs
… src/ops/delete.rs
… src/ops/mod.rs
… src/ops/read.rs
… src/ops/stat.rs
… src/ops/write.rs
… src/readers/callback.rs
… src/readers/mod.rs
… src/readers/seekable.rs
… src/readers/stream.rs
… src/services/fs.rs
… src/services/s3.rs
… tests/it/io.rs
… tests/it/layer.rs
… tests/it/main.rs
… tests/it/readers.rs
… tests/it/services/behavior.rs
… tests/it/wraps.rs