Signed-off-by: Xuanwo <github@xuanwo.io>

I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

See RFC-0041: https://github.com/datafuselabs/opendal/pull/41

With this refactor, we improve the performance on fs a lot:

- On my PC: the throughput increased from 400MB/s to 800MB/s
- On benchmark on databend sides: time for the same query will drop from 0.8s to 0.37s

## Unresolved problems

- We need to support construct `Object` via external information
- Maybe find a better name for `stateful_read`.